Make a reusable component: ListPage.tsx
